chore: move liquidation_math_tests
