+++ state-chain/pallets/cf-lending-pools/src/general_lending.rs
@@ -1301,142 +1301,6 @@ fn compute_per_asset_liquidation_estimates(
 		.collect()
 }
 
-#[cfg(test)]
-mod liquidation_math_tests {
-	use super::*;
-
-	const ETH: Asset = Asset::Eth;
-	const BTC: Asset = Asset::Btc;
-	const SOL: Asset = Asset::Sol;
-	const USDC: Asset = Asset::Usdc;
-
-	fn bps_to_permill(bps: BasisPoints) -> Permill {
-		Permill::from_rational(bps.min(ONE_AS_BASIS_POINTS) as u32, ONE_AS_BASIS_POINTS as u32)
-	}
-
-	fn estimates(
-		total_owed_usd: AssetAmount,
-		slippage_bps: BasisPoints,
-		positions: &[(Asset, AssetAmount, AssetAmount)],
-	) -> Vec<(Asset, AssetAmount)> {
-		compute_per_asset_liquidation_estimates(
-			required_collateral_with_buffer(total_owed_usd, bps_to_permill(slippage_bps)),
-			positions,
-		)
-	}
-
-	#[test]
-	fn required_collateral_with_no_buffer_equals_owed() {
-		assert_eq!(required_collateral_with_buffer(1_000, Permill::zero()), 1_000);
-	}
-
-	#[test]
-	fn required_collateral_with_buffer_inflates_with_ceil_rounding() {
-		// 1_000 / (1 - 0.005) = 1_005.025... -> 1_006 (ceil).
-		assert_eq!(required_collateral_with_buffer(1_000, bps_to_permill(50)), 1_006);
-		// 1_000 / (1 - 0.05) = 1_052.63... -> 1_053 (ceil).
-		assert_eq!(required_collateral_with_buffer(1_000, bps_to_permill(500)), 1_053);
-		// 65 bps combined buffer: 1_000 / (1 - 0.0065) = 1_006.54... -> 1_007 (ceil).
-		assert_eq!(required_collateral_with_buffer(1_000, bps_to_permill(65)), 1_007);
-	}
-
-	#[test]
-	fn required_collateral_saturates_when_buffer_is_total_loss() {
-		// A 100% buffer means any swap could return zero, so no finite amount is enough.
-		assert_eq!(required_collateral_with_buffer(1_000, Permill::one()), AssetAmount::MAX);
-	}
-
-	#[test]
-	fn empty_positions_give_no_estimates() {
-		assert!(estimates(1_000, 50, &[]).is_empty());
-	}
-
-	#[test]
-	fn single_asset_with_excess_takes_only_required_with_buffer() {
-		// Owed 1_000 USD, slippage 0.5%, available 10_000 USD worth of ETH (10_000 ETH @ $1).
-		// Required = ceil(1_000 / 0.995) = 1_006. ETH price is 1, so amount equals USD.
-		assert_eq!(estimates(1_000, 50, &[(ETH, 10_000, 10_000)]), vec![(ETH, 1_006)]);
-	}
-
-	#[test]
-	fn single_asset_with_deficit_takes_everything() {
-		// Owed 1_000 USD, slippage 0.5% → required 1_006 USD. Only 800 USD available.
-		assert_eq!(estimates(1_000, 50, &[(ETH, 800, 800)]), vec![(ETH, 800)]);
-	}
-
-	#[test]
-	fn at_required_threshold_takes_all() {
-		// Exactly enough collateral (in USD) to cover the slippage-buffered owed amount.
-		assert_eq!(estimates(1_000, 50, &[(ETH, 1_006, 1_006)]), vec![(ETH, 1_006)]);
-	}
-
-	#[test]
-	fn zero_owed_takes_nothing() {
-		assert!(estimates(0, 50, &[(ETH, 1_000, 1_000)]).is_empty());
-	}
-
-	#[test]
-	fn two_assets_with_equal_usd_split_evenly() {
-		// Owed 1_000 USD, slippage 0%, available: 100 ETH @ $5 and 50 BTC @ $10 = $500 each.
-		// target = 1_000, total_avail = 1_000 → take everything.
-		assert_eq!(
-			estimates(1_000, 0, &[(ETH, 100, 500), (BTC, 50, 500)]),
-			vec![(ETH, 100), (BTC, 50)]
-		);
-	}
-
-	#[test]
-	fn two_assets_drawn_proportionally_to_usd_share() {
-		// ETH: 100 units @ $1 = $100 USD. SOL: 200 units @ $5 = $1000 USD. Total = $1100.
-		// Owed 550 USD, slippage 0% → target = 550. ETH share = 100/1100 ≈ 9.09%, SOL = 90.91%.
-		// ETH take = ceil(100 * 550 / 1100) = 50. SOL take = ceil(200 * 550 / 1100) = 100.
-		assert_eq!(
-			estimates(550, 0, &[(ETH, 100, 100), (SOL, 200, 1_000)]),
-			vec![(ETH, 50), (SOL, 100)]
-		);
-	}
-
-	#[test]
-	fn three_assets_proportional_split_with_slippage_buffer() {
-		// Total available: 1_000 ETH ($2_000) + 500 SOL ($500) + 1_000_000 USDC ($1_000_000) =
-		// $1_002_500. Owed 500_000 USD, slippage 0.5% → required = ceil(500_000 / 0.995) =
-		// 502_513. Per-asset: ETH = ceil(1_000 * 502_513 / 1_002_500) ≈ 502, SOL = ceil(500 *
-		// 502_513 / 1_002_500) ≈ 251, USDC = ceil(1_000_000 * 502_513 / 1_002_500) ≈ 501_260.
-		assert_eq!(
-			estimates(
-				500_000,
-				50,
-				&[(ETH, 1_000, 2_000), (SOL, 500, 500), (USDC, 1_000_000, 1_000_000)],
-			),
-			vec![(ETH, 502), (SOL, 251), (USDC, 501_260)]
-		);
-	}
-
-	#[test]
-	fn rounds_up_so_collected_usd_meets_target() {
-		// ETH and SOL each contribute $50 USD with awkward unit counts that force rounding.
-		// Owed 33 USD, slippage 0% → target 33 USD, total avail 100 USD. Each share = 33/2 USD
-		// rounded up. ETH: ceil(7 * 33 / 100) = ceil(2.31) = 3. SOL: ceil(13 * 33 / 100) =
-		// ceil(4.29) = 5. The sum (8 + leftover from rounding) covers the target.
-		assert_eq!(estimates(33, 0, &[(ETH, 7, 50), (SOL, 13, 50)]), vec![(ETH, 3), (SOL, 5)]);
-	}
-
-	#[test]
-	fn deficit_in_some_assets_still_takes_all_when_total_required_exceeds_available() {
-		// Required 100_000 USD with 0% slippage but only $300 + $200 = $500 available → take all.
-		assert_eq!(
-			estimates(100_000, 0, &[(ETH, 30, 300), (SOL, 100, 200)],),
-			vec![(ETH, 30), (SOL, 100)]
-		);
-	}
-
-	#[test]
-	fn extreme_slippage_forces_take_all() {
-		// 99.5% slippage means required ≈ 200x owed → almost certainly exceeds available.
-		assert_eq!(estimates(100, 9_950, &[(ETH, 50, 50)]), vec![(ETH, 50)]);
-	}
-}
-
 /// Check collateralisation ratio (triggering/aborting liquidations if necessary) and
 /// periodically swap collected fees into each pool's desired asset.
 pub fn lending_upkeep<T: Config>(current_block: BlockNumberFor<T>) -> Weight {
